renamed old inbounds test and added a new no bounds test
